+++ factorialhr_analysis/pages/__init__.py
@@ -1,0 +1,4 @@
+from factorialhr_analysis.pages.index import index_page
+from factorialhr_analysis.pages.login import login_page
+
+__all__ = ['index_page', 'login_page']