+++ factorialhr_analysis/routes.py
@@ -1,0 +1,4 @@
+"""Routes for the application."""
+
+INDEX = '/'
+LOGIN_ROUTE = '/login'